refactor: [UF-301] fetch 제거 및 도메인에 구애받지 않는 순수 UI 컴포넌트로 분리 #330
